z-wave-stack-docker-23: legal: Add SDPX headers for SBOM and comply reuse spec
The copyright is set to the entity that provided the code, in future
more entities (eg: ZWA) can be added if substantial contributions are
landing in related files.

This legal effort is done to my best knowledge, with the help or reuse tool.
If any claim is not correct please file issues to the upstream project
it will be fixed accordingly, (check related URL in footer).

Origin: https://github.com/Z-Wave-Alliance/z-wave-stack-docker/pull/23
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack/issues/116
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>